requirements.txt
Django==4.2.7
python-dotenv==0.21.0
whitenoise==6.2.0
gunicorn==22.0.0
psycopg2-binary==2.9.5
django-heroku==0.3.1
requests==2.28.1
django-cors-headers==3.14.0
setuptoolsdjango-unfold==0.77.1
